pyproject.toml
[project]
name = "multicampus-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dotenv>=0.9.9",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langchain-openai>=0.3.28",
    "openai>=1.98.0",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "tavily-python>=0.7.10",
    "tiktoken>=0.9.0",
]
